services/ingestion/requirements.txt
requests==2.28.2
SQLAlchemy==1.4.47
psycopg2-binary==2.9.5
python-dotenv==1.0.0
kafka-python==2.0.2
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
Flask==2.0.3
Werkzeug==2.0.3
pyrate-limiter==2.10.0
pybreaker==1.0.2
prometheus-client==0.16.0
fastapi==0.95.0
uvicorn==0.21.1